+++ src/main/java/roomescape/presentation/controller/ReservationAdminController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.ReservationAdminRegisterDto;
-import roomescape.model.LoginMember;
+import roomescape.dto.LoginMember;
 import roomescape.service.ReservationAdminService;
 
 @RestController
