+++ src/main/java/roomescape/presentation/resvoler/LoginMemberArgumentResolver.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.model.LoginMember;
+import roomescape.dto.LoginMember;
 import roomescape.model.Member;
 import roomescape.presentation.support.CookieUtils;
 import roomescape.service.AuthService;
